+++ packages/lib/src/action/index.ts
@@ -9,11 +9,5 @@ export {
 export { HookAction, isHookAction } from "./hookActions"
 export { ActionMiddleware, ActionMiddlewareDisposer, addActionMiddleware } from "./middleware"
 export { isModelAction, modelAction } from "./modelAction"
-export {
-  asModelFlow,
-  FlowFunction,
-  FlowFunctionToPromiseFunction,
-  isModelFlow,
-  modelFlow,
-} from "./modelFlow"
+export { isModelFlow, modelFlow, _async, _await } from "./modelFlow"
 export { runUnprotected } from "./protection"