+++ packages/lib/src/index.ts
@@ -13,11 +13,3 @@ export * from "./tweaker"
 export * from "./typeChecking"
 export { MobxKeystoneError } from "./utils"
 export * from "./wrappers"
-
-declare global {
-  // used to make yield* work over promises (for flows)
-  // which is actually polyfilled in modelFlow.ts
-  interface Promise<T> {
-    [Symbol.iterator](): Generator<Promise<T>, T, unknown>
-  }
-}