+++ packages/lib/test/redux/connectReduxDevTools.test.ts
@@ -1,6 +1,5 @@
 import {
   applySnapshot,
-  asModelFlow,
   connectReduxDevTools,
   model,
   Model,
@@ -8,6 +7,8 @@ import {
   modelFlow,
   modelSnapshotOutWithMetadata,
   prop,
+  _async,
+  _await,
 } from "../../src"
 
 jest.useRealTimers()
@@ -69,37 +70,37 @@ class M extends Model({
 
   private *_setXAsync() {
     this.x = "setXAsync +0"
-    yield* waitAsync(20)
+    yield* _await(waitAsync(20))
     this.x = "setXAsync +20"
   }
 
   @modelFlow
-  setXAsync = asModelFlow(this._setXAsync)
+  setXAsync = _async(this._setXAsync)
 
   private *_setXAsyncWithEmptyFirstPart() {
-    yield* waitAsync(20)
+    yield* _await(waitAsync(20))
     this.x = "setXAsyncWithEmptyFirstPart +20"
   }
 
   @modelFlow
-  setXAsyncWithEmptyFirstPart = asModelFlow(this._setXAsyncWithEmptyFirstPart)
+  setXAsyncWithEmptyFirstPart = _async(this._setXAsyncWithEmptyFirstPart)
 
   private *_setXAsyncThrowSync() {
     this.x = "setXAsyncThrowSync +0"
-    yield* waitAsync(20)
+    yield* _await(waitAsync(20))
     throw new Error("setXAsyncThrowSync +20")
   }
 
   @modelFlow
-  setXAsyncThrowSync = asModelFlow(this._setXAsyncThrowSync)
+  setXAsyncThrowSync = _async(this._setXAsyncThrowSync)
 
   private *_setXAsyncThrowAsync() {
     this.x = "setXAsyncThrowAsync +0"
-    yield* waitAsyncReject(20)
+    yield* _await(waitAsyncReject(20))
   }
 
   @modelFlow
-  setXAsyncThrowAsync = asModelFlow(this._setXAsyncThrowAsync)
+  setXAsyncThrowAsync = _async(this._setXAsyncThrowAsync)
 
   @modelAction
   setY() {
@@ -114,29 +115,29 @@ class M extends Model({
 
   private *_setYAsync() {
     this.y = "setYAsync +0"
-    yield* waitAsync(50)
+    yield* _await(waitAsync(50))
     this.y = "setYAsync +50"
   }
 
   @modelFlow
-  setYAsync = asModelFlow(this._setYAsync)
+  setYAsync = _async(this._setYAsync)
 
   private *_setYAsyncThrowSync() {
     this.y = "setYAsyncThrowSync +0"
-    yield* waitAsync(50)
+    yield* _await(waitAsync(50))
     throw new Error("setYAsyncThrowSync +50")
   }
 
   @modelFlow
-  setYAsyncThrowSync = asModelFlow(this._setYAsyncThrowSync)
+  setYAsyncThrowSync = _async(this._setYAsyncThrowSync)
 
   private *_setYAsyncThrowAsync() {
     this.y = "setYAsyncThrowAsync +0"
-    yield* waitAsyncReject(50)
+    yield* _await(waitAsyncReject(50))
   }
 
   @modelFlow
-  setYAsyncThrowAsync = asModelFlow(this._setYAsyncThrowAsync)
+  setYAsyncThrowAsync = _async(this._setYAsyncThrowAsync)
 
   @modelAction
   setXY() {
@@ -148,27 +149,27 @@ class M extends Model({
 
   private *_setXYAsync() {
     this.x = "setXYAsync starts"
-    yield* this.setXAsync()
-    yield* this.setYAsync()
+    yield* _await(this.setXAsync())
+    yield* _await(this.setYAsync())
     this.x = "setXYAsync ends"
   }
 
   @modelFlow
-  setXYAsync = asModelFlow(this._setXYAsync)
+  setXYAsync = _async(this._setXYAsync)
 
   private *_setXYAsyncThrowSync() {
-    yield* this.setXAsyncThrowSync()
+    yield* _await(this.setXAsyncThrowSync())
   }
 
   @modelFlow
-  setXYAsyncThrowSync = asModelFlow(this._setXYAsyncThrowSync)
+  setXYAsyncThrowSync = _async(this._setXYAsyncThrowSync)
 
   private *_setXYAsyncThrowAsync() {
-    yield* this.setXAsyncThrowAsync()
+    yield* _await(this.setXAsyncThrowAsync())
   }
 
   @modelFlow
-  setXYAsyncThrowAsync = asModelFlow(this._setXYAsyncThrowAsync)
+  setXYAsyncThrowAsync = _async(this._setXYAsyncThrowAsync)
 }
 
 let m = new M({})